requirements/test.txt
# Test dependencies go here.
-r base.txt
coverage==3.6
django-discover-runner==1.0
django-nose==1.2
factory-boy==2.1.2
selenium==2.35.0
